Refactor package name to com.xcv58.graphql
